pyproject.toml
[tool.poetry]
name = "coral"
version = "0.1.0"
description = ""
authors = ["pavelborisov <pavel.borisofff@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
fastapi = "^0.73.0"
uvicorn = "^0.17.3"
httpx = "^0.22.0"
motor = "^2.5.1"
python-multipart = "^0.0.5"
dnspython = "^2.2.0"
python-dotenv = "^0.19.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
